Find bug ( test/MAP/OK_kusano/a.cub )
