+++ lib/debug/Makefile
@@ -6,7 +6,7 @@
 #    By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/31 21:09:57 by kamitsui          #+#    #+#              #
-#    Updated: 2024/11/18 16:40:15 by kamitsui         ###   ########.fr        #
+#    Updated: 2024/12/17 03:16:40 by kamitsui         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -46,6 +46,8 @@ endif
 
 # Source files
 SRCS = \
+	   put_visited.c \
+	   debug_parse_fc.c \
 	   debug_map_data.c \
 	   debug_player.c \
 	   debug_ray.c \
