+++ lib/debug/includes/debug_cub3d.h
@@ -6,7 +6,7 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/31 20:42:19 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/12/16 23:24:54 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/12/17 03:16:17 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include "type_cub3d.h"
 
 # define LOG_FILE "debug.log"
-# define IS_DEBUG 1
+# define IS_DEBUG 0
 # define DEBUG_COLOR "\n\x1B[100m\x1B[37m"
 # define RESET_COLOR "\x1B[0m\n"
 
@@ -26,13 +26,13 @@ void	debug_element_type(int fd, const char *str, const char *msg);
 void	debug_texture(char *file, t_texture texture,
 			const char *msg);
 void	debug_img_tex(int fd, t_img img, int i, const char *msg);
+void	debug_tex_info(int fd, t_info tex_info);
+void	debug_parse_fc(int fd, int color[2], const char *msg);
 void	debug_map_data(t_map map, const char *msg);
 void	debug_player(int fd, t_player player, const char *msg);
 void	debug_get_rgb_color(int fd,
 			t_type_fc type, int rgb[3], const char *msg);
-
-// parse cubfile
-void	debug_tex_info(int fd, t_info tex_info);
+void	put_visited(int fd, bool **visited, t_map *map);
 
 // render frame
 void	debug_moved_player(t_game *game);
