+++ map/test.cub
@@ -1,14 +1,34 @@
 
 
-NO ./texture/bookshelf.xpm
-C 225,30,0
 
-EA ./texture/42_logo.xpm
-SO ./texture/kamecci.xpm
+NO ./TEXTURE/kita.xpm						
 
-WE ./texture/tails.xpm
+WE ./TEXTURE/nishi.xpm     				    	
 
+SO ./TEXTURE/minami.xpm							
 
 
 
-F            220,100,0
+EA ./TEXTURE/higashi.xpm                          																
+
+C 12,152,1          
+
+
+F 1,8,77
+     11                   
+    1001
+   100001             
+  100N0001
+ 1000100001
+10001 100001
+ 1000100001               
+  10000001       
+   100001                
+    1001
+     11     
+
+
+
+
+
+
