+++ srcs/init_utils/parse_cubfile_utils/parse_map_utils/check_enclosed_by_walls.c
@@ -6,24 +6,11 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 13:51:35 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/12/17 01:11:11 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/12/17 03:19:08 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
-
-// debug
-void	put_visited(int fd, bool **visited, t_map *map)
-{
-	for (int r = 0; r < map->height; r++)
-	{
-		for (int c = 0; c < map->width; c++)
-		{
-			ft_dprintf(fd, "%d", visited[r][c]);
-		}
-		ft_dprintf(fd, "\n");
-	}
-}
 
 static void	put_error_msg_is_not_map_enclosed(void)
 {
@@ -43,14 +30,8 @@ static bool	is_enclosed_on_remaining_area(
 		while (x < map->width)
 		{
 			if (map->data[y][x] == '0' && visited[y][x] == false)
-			{
-				ft_dprintf(map->debug.fd, 
-						"visited[%d][%d]={%d} data={%d}\n", y,x,visited[y][x],map->data[y][x]);
 				if (flood_fill(map, x, y, (bool **)visited) != true)
 					return (false);
-				ft_dprintf(map->debug.fd, 
-						"return={%d} 1:true\n", true);
-			}
 			x++;
 		}
 		y++;
@@ -109,7 +90,6 @@ int	check_enclosed_by_walls(const char *line, t_parse *parse)
 	t_map	*map;
 	t_bool	is_surrounded;
 
-	int	fd = parse->game->debug.fd;// for debug
 	visited = init_visited(&parse->game->map);
 	if (visited == NULL)
 		return (EXIT_FAILURE);
@@ -117,24 +97,17 @@ int	check_enclosed_by_walls(const char *line, t_parse *parse)
 	start_y = parse->player_grid.y;
 	map = &parse->game->map;
 	is_surrounded = ENUM_TRUE;
-	put_visited(fd, visited, map);// debug
 	is_surrounded = flood_fill(map, start_x, start_y, visited);
-	put_visited(fd, visited, map);// debug
-	ft_dprintf(fd, "enclosed[%d]\n", is_surrounded);
 	if (process_false_or_error(is_surrounded) != ENUM_TRUE)
 		return (EXIT_FAILURE);
-	// clear_visited();
 	is_surrounded = is_enclosed_on_remaining_area(map, visited);
-	put_visited(fd, visited, map);// debug
-	ft_dprintf(fd, "enclosed[%d]\n", is_surrounded);
 	if (process_false_or_error(is_surrounded) != ENUM_TRUE)
 		return (EXIT_FAILURE);
-	//exit(0);
-	//ft_printf("here?\n");
 	free_visited(visited, (size_t)map->height);
 	(void)line;
 	return (EXIT_SUCCESS);
 }
 // debug code
 //printf("-- End flood_fill()\n");//debug
-//put_visited(parse->game->debug.fd, visited, map);
+//int	fd = parse->game->debug.fd;// for debug
+//put_visited(fd, visited, map);
