+++ srcs/init_utils/parse_cubfile_utils/parse_map_utils/flood_fill.c
@@ -6,7 +6,7 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/01 02:42:15 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/12/16 23:50:00 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/12/17 03:20:15 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,11 +76,6 @@ int	check_grid_is_not_visited(t_point current, t_map *map,
 			*is_surrounded = ENUM_FALSE;
 		return (CONTINUE);
 	}
-	int	x = current.x;
-	int	y = current.y;
-	ft_dprintf(map->debug.fd, "x[%d] y[%d] map->width[%d] map->height[%d] ",
-			x, y, map->width, map->height);
-	ft_dprintf(map->debug.fd, "visited[%d] data['%c']\n", visited[y][x], map->data[y][x]);
 	return (EXIT_SUCCESS);
 }
 
@@ -107,7 +102,6 @@ t_bool	flood_fill(t_map *map, int start_x, int start_y, bool **visited)
 	t_bool		is_surrounded;
 	t_point		current;
 
-	ft_dprintf(map->debug.fd, ">> flood_fill <<\n");
 	if (initialize_for_flood_fill(&stack, start_x, start_y) != EXIT_SUCCESS)
 		return (ENUM_ERROR);
 	is_surrounded = ENUM_TRUE;
